android implementation fixes

+++ defunityads/src/ios/DefUnityAdsIOS.mm
@@ -38,29 +38,29 @@ UIView *bannerView;
 @implementation DefUnityAdsBannerDelegate
 
 -(void) unityAdsBannerDidClick: (NSString *) placementId {
-  DefUnityCallback_add_to_queue((int)TYPE_BANNER,(char*)"placementId", (char*)[placementId UTF8String], (char*)"event", BANNER_EVENT_DID_CLICK);
+  DefUnityCallback_add_to_queue((int)TYPE_BANNER,(char*)"placementId", (char*)[placementId UTF8String], (char*)"event", (int)BANNER_EVENT_DID_CLICK);
 }
 
 -(void) unityAdsBannerDidError: (NSString *) message {
-  DefUnityCallback_add_to_queue((int)TYPE_BANNER,(char*)"message", (char*)[message UTF8String], (char*)"event", BANNER_EVENT_DID_ERROR);
+  DefUnityCallback_add_to_queue((int)TYPE_BANNER,(char*)"message", (char*)[message UTF8String], (char*)"event", (int)BANNER_EVENT_DID_ERROR);
 }
 
 -(void) unityAdsBannerDidHide: (NSString *) placementId {
-  DefUnityCallback_add_to_queue((int)TYPE_BANNER,(char*)"placementId", (char*)[placementId UTF8String], (char*)"event", BANNER_EVENT_DID_HIDE);
+  DefUnityCallback_add_to_queue((int)TYPE_BANNER,(char*)"placementId", (char*)[placementId UTF8String], (char*)"event", (int)BANNER_EVENT_DID_HIDE);
 }
 
 -(void) unityAdsBannerDidShow: (NSString *) placementId {
-  DefUnityCallback_add_to_queue((int)TYPE_BANNER,(char*)"placementId", (char*)[placementId UTF8String], (char*)"event", BANNER_EVENT_DID_SHOW);
+  DefUnityCallback_add_to_queue((int)TYPE_BANNER,(char*)"placementId", (char*)[placementId UTF8String], (char*)"event", (int)BANNER_EVENT_DID_SHOW);
 }
 
 -(void) unityAdsBannerDidLoad: (NSString *) placementId view: (UIView *) view {
   bannerView = view;
-  DefUnityCallback_add_to_queue((int)TYPE_BANNER,(char*)"placementId", (char*)[placementId UTF8String], (char*)"event", BANNER_EVENT_DID_LOAD);
+  DefUnityCallback_add_to_queue((int)TYPE_BANNER,(char*)"placementId", (char*)[placementId UTF8String], (char*)"event", (int)BANNER_EVENT_DID_LOAD);
 }
 
 -(void) unityAdsBannerDidUnload: (NSString *) placementId {
   bannerView = nil;
-  DefUnityCallback_add_to_queue((int)TYPE_BANNER,(char*)"placementId", (char*)[placementId UTF8String], (char*)"event", BANNER_EVENT_DID_UNLOAD);
+  DefUnityCallback_add_to_queue((int)TYPE_BANNER,(char*)"placementId", (char*)[placementId UTF8String], (char*)"event", (int)BANNER_EVENT_DID_UNLOAD);
 }
 @end
